Move mapped-comments to root level
