+++ src/ggrc/assets/assets.yaml
@@ -16,11 +16,11 @@ dashboard-js-template-files:
   - components/assessment/mapped-objects/mapped-control-related-objects.mustache
   - components/assessment/mapped-objects/mapped-controls.mustache
   - components/assessment/mapped-objects/mapped-related-information.mustache
-  - components/assessment/mapped-objects/mapped-comments.mustache
   - components/assessment/custom-attributes.mustache
   - components/gca-controls/gca-controls.mustache
   - components/comment/comment-add-form.mustache
   - components/comment/comment-input.mustache
+  - components/comment/mapped-comments.mustache
   - components/object-list/object-list.mustache
   - components/object-popover/object-popover.mustache
   - components/object-popover/related-assessment-popover.mustache
