+++ src/ggrc/assets/javascripts/components/assessment/info-pane/info-pane.js
@@ -9,7 +9,8 @@ import '../assessment-custom-attributes';
 import '../assessment-people';
 import '../assessment-object-type-dropdown';
 import '../mapped-objects/mapped-related-information';
-import '../mapped-objects/mapped-comments';
+import '../../comment/comment-add-form';
+import '../../comment/mapped-comments';
 import '../mapped-objects/mapped-controls';
 import '../../assessment/map-button-using-assessment-type';
 import '../../ca-object/ca-object-modal-content';
