+++ src/ggrc/assets/javascripts/components/comment/mapped-comments.js
@@ -1,18 +1,17 @@
-/*!
+/*
  Copyright (C) 2017 Google Inc., authors, and contributors
  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
 
-import '../../object-list-item/comment-list-item';
-import '../../object-list/object-list';
+import '../object-list-item/comment-list-item';
+import '../object-list/object-list';
 
 (function (can, GGRC) {
   'use strict';
 
   var tpl = can.view(GGRC.mustache_path +
-    '/components/assessment/mapped-objects/' +
-    'mapped-comments.mustache');
-  var tag = 'assessment-mapped-comments';
+    '/components/comment/mapped-comments.mustache');
+  var tag = 'mapped-comments';
   /**
    * Assessment specific mapped controls view component
    */
@@ -23,10 +22,10 @@ import '../../object-list/object-list';
       define: {
         noItemsText: {
           type: 'string',
-          value: ''
-        }
+          value: '',
+        },
       },
-      mappedItems: []
-    }
+      mappedItems: [],
+    },
   });
 })(window.can, window.GGRC);
