+++ src/ggrc/assets/mustache/components/assessment/info-pane/info-pane.mustache
@@ -270,7 +270,7 @@
                                     (before-create)="addItems(%event, 'comments')"
                     ></comment-add-form>
                 {{/unless}}
-                <assessment-mapped-comments {mapped-items}="comments"></assessment-mapped-comments>
+                <mapped-comments {mapped-items}="comments"></mapped-comments>
               </div>
             </div>
         </div>
