+++ src/ggrc/assets/stylesheets/components/comment/_comment-list.scss
@@ -2,7 +2,7 @@
  * Copyright (C) 2017 Google Inc.
  * Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
  */
-assessment-mapped-comments {
+mapped-comments {
   overflow-y: auto;
   overflow-x: hidden;
   display: flex;
